+++ Software/Photonfirmware/AnimatronicEyesTest/src/TPPAnimateServo.h
@@ -41,7 +41,7 @@ class TPP_AnimateServo{
         TPP_AnimateServo();
         void begin(int servoNum, int postion);
         void process(); // called every time in the loop to keep the eyes moving
-        int moveTo (int newX, int speed);
+        int moveTo (int newX, float speed);
 
     private:
